Module Lexicographic.Hash_queue

type key = t

include Core.Container.S1_phantom with type ('data, 'key) t := ('key, 'data) Core.Hash_queue.t

include sig ... end

val length : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> int

val is_empty : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> bool

val iter : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> f:('a -> unit) @ local -> unit

iter must allow exceptions raised in f to escape, terminating the iteration cleanly. The same holds for all functions below taking an f.

val exists : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> f:('a -> bool) @ local -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> f:('a -> bool) @ local -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> f:('a -> bool) @ local -> int

Returns the number of elements for which the provided function evaluates to true.

val find : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> f:('a -> bool) @ local -> 'a Option.t

Returns as an option the first element for which f evaluates to true.

val to_list : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> 'a list

val min_elt : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> compare:('a -> ('a -> int) @ local) @ local -> 'a Option.t

Returns a min (resp. max) element from the collection using the provided compare function. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold. Returns None iff the collection is empty.

val max_elt : 'a 'p1 'p2. ('p1, 'a) Core.Hash_queue.t -> compare:('a -> ('a -> int) @ local) @ local -> 'a Option.t

val sum : 'a 'sum 'p1 'p2. (module Base.Container.Summable with type t = 'sum) -> ('p1, 'a) Core.Hash_queue.t -> f:('a -> 'sum) @ local -> 'sum

Returns the sum of f i for all i in the container. The order in which the elements will be summed is unspecified.

val iter_until : 'a 'p1 'p2 'final. ('p1, 'a) Core.Hash_queue.t -> f:('a -> (unit, 'final) Base.Container.Continue_or_stop.t) @ local -> (finish:(unit -> 'final) @ local -> 'final) @ local

iter_until t ~f ~finish is a short-circuiting version of iter. If f returns Stop x the computation ceases and returns x. If f always returns Continue () the final result is computed by finish.

val fold : 'a 'p1 'p2 'acc. ('p1, 'a) Core.Hash_queue.t -> init:'acc -> f:('acc -> ('a -> 'acc) @ local) @ local -> 'acc

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : 'a 'p1 'p2 'acc 'e. ('p1, 'a) Core.Hash_queue.t -> init:'acc -> f:('acc -> ('a -> ('acc, 'e) Result.t) @ local) @ local -> ('acc, 'e) Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val find_map : 'a 'p1 'p2 'b. ('p1, 'a) Core.Hash_queue.t -> f:('a -> 'b Option.t) @ local -> 'b Option.t

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val fold_until : 'a 'p1 'p2 'acc 'final. ('p1, 'a) Core.Hash_queue.t -> init:'acc -> f: ('acc -> ('a -> ('acc, 'final) Base.Container.Continue_or_stop.t) @ local) @ local -> (finish:('acc -> 'final) @ local -> 'final) @ local

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

  type maybe_negative =
    | Found_negative of int
    | All_nonnegative of { sum : int }

  (** [first_neg_or_sum list] returns the first negative number in [list], if any,
      otherwise returns the sum of the list. *)
  let first_neg_or_sum =
    List.fold_until ~init:0
      ~f:(fun sum x ->
        if x < 0
        then Stop (Found_negative x)
        else Continue (sum + x))
      ~finish:(fun sum -> All_nonnegative { sum })
  ;;

  let x = first_neg_or_sum [1; 2; 3; 4; 5]
  val x : maybe_negative = All_nonnegative {sum = 15}

  let y = first_neg_or_sum [1; 2; -3; 4; 5]
  val y : maybe_negative = Found_negative -3

val to_array : 'a 'p. ('p, 'a) Core.Hash_queue.t -> 'a array

invariant t checks the invariants of the queue.

val invariant : ('key, 'data) Core.Hash_queue.t -> Base.Unit.t

val create : ?growth_allowed:Base.Bool.t -> ?size:Base.Int.t -> Base.Unit.t -> (key, 'data) Core.Hash_queue.t

create () returns an empty queue. The arguments growth_allowed and size are referring to the underlying hashtable.

parameter growth_allowed
defaults to true

parameter size
initial size -- default to 16

val clear : ('key, 'data) Core.Hash_queue.t -> Base.Unit.t

Clears the queue.

val copy : ('key, 'data) Core.Hash_queue.t -> ('key, 'data) Core.Hash_queue.t

Makes a fresh copy of the queue with identical contents to the original.

Finding elements

val mem : ('key, 'data) Core.Hash_queue.t -> 'key -> Base.Bool.t

mem q k returns true iff there is some (k, v) in the queue.

val lookup : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data Base.Option.t

lookup t k returns the value of the key-value pair in the queue with key k, if there is one.

val lookup_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data

Adding, removing, and replacing elements

Note that even the non-*_exn versions can raise, but only if there is an ongoing iteration.

val enqueue : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> 'key -> 'data -> [ `Ok | `Key_already_present ]

enqueue t back_or_front k v adds the key-value pair (k, v) to the front or back of the queue, returning `Ok if the pair was added, or `Key_already_present if there is already a (k, v') in the queue.

val enqueue_exn : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> 'key -> 'data -> Base.Unit.t

Like enqueue, but it raises in the `Key_already_present case

val enqueue_back : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> [ `Ok | `Key_already_present ]

See enqueue. enqueue_back t k v is the same as enqueue t `back k v

val enqueue_back_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> Base.Unit.t

See enqueue_exn. enqueue_back_exn t k v is the same as enqueue_exn t `back k v

val enqueue_front : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> [ `Ok | `Key_already_present ]

See enqueue. enqueue_front t k v is the same as enqueue t `front k v

val enqueue_front_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> Base.Unit.t

See enqueue_exn. enqueue_front_exn t k v is the same as enqueue_exn t `front k v

val lookup_and_move_to_back : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data Base.Option.t

lookup_and_move_to_back finds the key-value pair (k, v) and moves it to the back of the queue if it exists, otherwise returning None.

The _exn versions of these functions raise if key-value pair does not exist.

val lookup_and_move_to_back_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data

Like lookup_and_move_to_back, but raises instead of returning an option

val lookup_and_move_to_front : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data Base.Option.t

Like lookup_and_move_to_back, but moves element to the front of the queue

val lookup_and_move_to_front_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data

Like lookup_and_move_to_front, but raises instead of returning an option

val last : ('key, 'data) Core.Hash_queue.t -> 'data Base.Option.t

last t returns the last element of the queue, without removing it.

val last_with_key : ('key, 'data) Core.Hash_queue.t -> ('key * 'data) Base.Option.t

last_with_key t returns the last element of the queue and its key, without removing it.

val first : ('key, 'data) Core.Hash_queue.t -> 'data Base.Option.t

first t returns the front element of the queue, without removing it.

val first_with_key : ('key, 'data) Core.Hash_queue.t -> ('key * 'data) Base.Option.t

first_with_key t returns the front element of the queue and its key, without removing it.

val keys : ('key, 'data) Core.Hash_queue.t -> 'key Base.List.t

keys t returns the keys in the order of the queue.

val lookup_prev_with_key : ('key, 'data) Core.Hash_queue.t -> 'key -> ('key * 'data) Base.Option.t

Find and return the key-value pair directly before 'key in the queue

val lookup_next_with_key : ('key, 'data) Core.Hash_queue.t -> 'key -> ('key * 'data) Base.Option.t

Find and return the key-value directly after 'key in the queue

val to_alist : ('key, 'data) Core.Hash_queue.t -> ('key * 'data) Base.List.t

to_alist t returns the elements of the queue with their keys in the order of the queue.

val of_alist_exn : ?growth_allowed:Base.Bool.t -> ?size:Base.Int.t -> Base.Unit.t -> (key * 'data) Base.List.t -> (key, 'data) Core.Hash_queue.t

of_alist_exn t creates a queue from the given list, raising in the same situation as enqueue_exn

val dequeue : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> 'data Base.Option.t

dequeue t front_or_back returns the front or back element of the queue.

val dequeue_exn : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> 'data

Like dequeue, but it raises if the queue is empty.

val dequeue_back : ('key, 'data) Core.Hash_queue.t -> 'data Base.Option.t

dequeue_back t returns the back element of the queue.

val dequeue_back_exn : ('key, 'data) Core.Hash_queue.t -> 'data

Like dequeue_back, but it raises if the queue is empty.

val dequeue_front : ('key, 'data) Core.Hash_queue.t -> 'data Base.Option.t

dequeue_front t returns the front element of the queue.

val dequeue_front_exn : ('key, 'data) Core.Hash_queue.t -> 'data

Like dequeue_front, but it raises if the queue is empty.

val dequeue_with_key : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> ('key * 'data) Base.Option.t

dequeue_with_key t returns the front or back element of the queue and its key.

val dequeue_with_key_exn : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> 'key * 'data

Like dequeue_with_key, but it raises if the queue is empty.

val dequeue_back_with_key : ('key, 'data) Core.Hash_queue.t -> ('key * 'data) Base.Option.t

dequeue_back_with_key t returns the back element of the queue and its key.

val dequeue_back_with_key_exn : ('key, 'data) Core.Hash_queue.t -> 'key * 'data

Like dequeue_back_with_key, but it raises if the queue is empty.

val dequeue_front_with_key : ('key, 'data) Core.Hash_queue.t -> ('key * 'data) Base.Option.t

dequeue_front_with_key t returns the front element of the queue and its key.

val dequeue_front_with_key_exn : ('key, 'data) Core.Hash_queue.t -> 'key * 'data

Like dequeue_front_with_key, but it raises if the queue is empty.

val dequeue_all : ('key, 'data) Core.Hash_queue.t -> f:('data -> Base.Unit.t) @ local -> Base.Unit.t

dequeue_all t ~f dequeues every element of the queue and applies f to each one. The dequeue order is from front to back.

val remove : ('key, 'data) Core.Hash_queue.t -> 'key -> [ `Ok | `No_such_key ]

remove q k removes the key-value pair with key k from the queue.

val remove_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> Base.Unit.t

val lookup_and_remove : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data Base.Option.t

like remove, but returns the removed element

val replace : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> [ `Ok | `No_such_key ]

replace q k v changes the value of key k in the queue to v.

val replace_or_enqueue : ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> 'key -> 'data -> Base.Unit.t

replace_or_enqueue q back_or_front k v changes the value of key k in the queue to v. If the key k does not exist in the queue, it is added to the front or back with the value v.

val replace_or_enqueue_front : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> Base.Unit.t

See replace_or_enqueue. replace_or_enqueue_front t k v is the same as replace_or_enqueue t `front k v

val replace_or_enqueue_back : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> Base.Unit.t

See replace_or_enqueue. replace_or_enqueue_back t k v is the same as replace_or_enqueue t `back k v

val replace_exn : ('key, 'data) Core.Hash_queue.t -> 'key -> 'data -> Base.Unit.t

val drop : ?n:Base.Int.t -> ('key, 'data) Core.Hash_queue.t -> [ `back | `front ] -> Base.Unit.t

drop ?n q back_or_front drops n elements (default 1) from the back or front of the queue. If the queue has fewer than n elements then it is cleared.

val drop_front : ?n:Base.Int.t -> ('key, 'data) Core.Hash_queue.t -> Base.Unit.t

Equivalent to drop ?n q `front.

val drop_back : ?n:Base.Int.t -> ('key, 'data) Core.Hash_queue.t -> Base.Unit.t

Equivalent to drop ?n q `back.

Iterating over elements

val iteri : ('key, 'data) Core.Hash_queue.t -> f:(key:'key -> (data:'data -> Base.Unit.t) @ local) @ local -> Base.Unit.t

iter t ~f applies f to each key and element of the queue.

val foldi : ('key, 'data) Core.Hash_queue.t -> init:'acc -> f:('acc -> (key:'key -> (data:'data -> 'acc) @ local) @ local) @ local -> 'acc

val findi : ('key, 'data) Core.Hash_queue.t -> f:(key:'key -> data:'data -> Base.Bool.t) -> ('key * 'data) Base.Option.t

type 'data t = (key, 'data) Core.Hash_queue.t

val sexp_of_t : ('data -> Sexplib0.Sexp.t) -> 'data t -> Sexplib0.Sexp.t